Move fixtures to conftest.py
Pytest fixtures in file with this name is automatically
available to all tests.
